[리팩터링] 코드 전반 수정
