+++ src/main/java/racingcar/domain/Car.java
@@ -22,12 +22,12 @@ public class Car {
         this.name = name;
     }
 
-    public String getName() {
-        return name;
-    }
-
     public void forward() {
         forwardCount++;
+    }
+
+    public String getName() {
+        return name;
     }
 
     public int getForwardCount() {
