+++ src/main/java/racingcar/domain/CarNames.java
@@ -21,20 +21,8 @@ public class CarNames {
     private List<String> carNames;
 
     public CarNames(String input) throws IllegalArgumentException {
-        validateHasBlank(input);
+        validateHasNotBlank(input);
         carNames = split(input);
-    }
-
-    private static void validateHasBlank(String input) {
-        if (input.contains(BLANK)) {
-            throw new IllegalArgumentException("공백은 허용되지 않습니다.");
-        }
-    }
-
-    private static void validateNameLength(String input) {
-        if (input.length() > NAME_LENGTH_LIMIT) {
-            throw new IllegalArgumentException("5자 이상 이름은 허용되지 않습니다.");
-        }
     }
 
     private static List<String> split(String input) {
@@ -43,6 +31,18 @@ public class CarNames {
             validateNameLength(carName);
         }
         return inputs;
+    }
+
+    private static void validateHasNotBlank(String input) {
+        if (input.contains(BLANK)) {
+            throw new IllegalArgumentException("공백은 허용되지 않습니다.");
+        }
+    }
+
+    private static void validateNameLength(String input) {
+        if (input.length() > NAME_LENGTH_LIMIT) {
+            throw new IllegalArgumentException(NAME_LENGTH_LIMIT + "자 이상 이름은 허용되지 않습니다.");
+        }
     }
 
     public List<String> getCarNames() {
