+++ src/main/java/racingcar/domain/TryCount.java
@@ -12,9 +12,10 @@ package racingcar.domain;
 
 public class TryCount {
 
-    private final static String EXCEPTION_MESSAGE = "1부터 1000까지의 숫자만 입력 가능합니다.";
     private final static int MINIMUM_NUMBER = 1;
     private final static int MAXIMUM_NUMBER = 1000;
+    private final static String EXCEPTION_MESSAGE =
+        MINIMUM_NUMBER + "부터" + MAXIMUM_NUMBER + "까지의 숫자만 입력 가능합니다.";
     private final int tryCount;
 
     public TryCount(String count) {
