+++ src/test/java/racingcar/domain/BehaviorJudgmentTest.java
@@ -14,6 +14,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.BehaviorJudgment;
 
 public class BehaviorJudgmentTest {
